Split SymbolExtractor language helpers (#4107)
