+++ changelog.d/unreleased/4107.internal.md
@@ -1,0 +1,18 @@
+---
+category: internal
+issues:
+  - 4107
+affected:
+  - src/CodeIndex/Indexer/Symbols/SymbolExtractor.cs
+  - src/CodeIndex/Indexer/Symbols/SymbolExtractor.Css.cs
+  - src/CodeIndex/Indexer/Symbols/SymbolExtractor.JavaScriptTypeScriptStyledFactory.cs
+  - docs/large-file-decomposition-plan.md
+---
+
+## English
+
+- **Started the SymbolExtractor language-helper decomposition path (#4107)** - moved the JS/TS styled-factory gate and CSS/Sass scanner helpers into focused partials so the shared core no longer owns those language-specific paths, without changing extraction semantics.
+
+## 日本語
+
+- **SymbolExtractor の言語別 helper 分割経路を開始しました (#4107)** - JS/TS styled-factory gate と CSS/Sass scanner helper を focused partial へ移し、共有 core がそれらの言語固有 path を持たないようにしました。抽出 semantics は変更していません。
